rewrite MT_Ntools
